% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\docType{data}
\name{box_cox_plus1}
\alias{box_cox_plus1}
\title{Box-cox transformation with a shift of 1 added to the data}
\format{An object of class \code{list} of length 2.}
\usage{
box_cox_plus1
}
\description{
\itemize{
\item{T} the transformation with arguments Y, the data, lambda the parameter, and boolean inverse to calculate inverse transformation.
\item{T_deriv} the transformation with arguments Y, the data, lambda the parameter.
}
}
\keyword{datasets}
