% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\docType{data}
\name{box_cox_shift}
\alias{box_cox_shift}
\title{Box-cox transformation of shifted variable}
\format{An object of class \code{list} of length 2.}
\usage{
box_cox_shift
}
\description{
\itemize{
\item{T} the transformation with arguments Y, the data, lambda the parameter, and boolean inverse to calculate inverse transformation. The parameter lambda has two real elements (1) the power and (2) the additive shift to the data.  
\item{T_deriv} the transformation with arguments Y, the data, lambda the parameter.
}
}
\keyword{datasets}
