% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrscale.R
\name{rrscale}
\alias{rrscale}
\title{a}
\usage{
rrscale(Y, trans_list, lims_list, opt_control = NULL, ncores = NULL,
  run_parallel = TRUE, verbose = TRUE, zeros = FALSE, seed = NULL)
}
\arguments{
\item{Y}{Data matrix to be transformed.}

\item{trans_list}{List of transformations to be considered. See function list_transformations.}

\item{lims_list}{List of optimization limits for each transformation from trans_list.}

\item{opt_control}{Optional optimization controlling parameters for DEoptim control argument.}

\item{ncores}{Number of cores to use if running parallel.}

\item{run_parallel}{TRUE/FALSE run column-wise in parallel.}

\item{verbose}{TRUE/FALSE save optimization output in local directory '.rrscale'}

\item{zeros}{How to deal with zeros in the data set. If set to FALSE the algorithm will fail if it encounters a zero. If set to 'adjust' then the zeros are replaced by 1E-10. If set to 'remove' then the zeros are set to NA.}

\item{seed}{Sets the seed before running any other analyses.}
}
\value{
A list of output:
opts: the optimization output for all transformation families and all columns
pars: the optimal parameters for each column for the optimal family
par_hat: the estimated optimal paramter
NT: the original data
RR: the robust-rescaled data
G: gaussianized data
Z: robust z-transformed data
O: data with outliers removed
transform: a function that takes the data and performs the RR transformation given the estimated values of the transformation parameter, mean, and sd of the data.
T: the optimal family
T_deriv: the derivative of the optimal family
T_name: name of the optimal family
alg_control: the parameters passed to the algorithm
}
\description{
a
}
