% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_na.R
\name{svdc}
\alias{svdc}
\title{The completed SVD}
\usage{
svdc(X, nu = NULL, nv = NULL)
}
\arguments{
\item{X}{the data matrix of which to calcluate the completed SVD.}

\item{nu}{the number of left singular vectors to calculate}

\item{nv}{the nubmer of right singular vectors to calculate}
}
\description{
This calculates right and left singular vectors of a data matrix possibly containing missing values.
}
\examples{
Y <- rnorm(10)\%*\%t(rnorm(10))
Y[1,1] <- NA
svdc.out <- svdc(Y)
}
