% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{winsor}
\alias{winsor}
\title{Winsorizes the data}
\usage{
winsor(x, fraction = 0.01)
}
\arguments{
\item{x}{the data.}

\item{fraction}{the top and bottom quantiles to cap.}
}
\description{
Winsorizes the data
}
\examples{
Y <- rlnorm(10)\%*\%t(rlnorm(10))
Yw <- winsor(Y,1E-2)
}
